move Dockerfile parser to docker.
